Broke test out into separate files
