+++ test/evaluation_test.rb
@@ -1,112 +1,6 @@
 require 'test_helper'
 
-FUNCTION_PATTERN = /\A\s*function\s*\(.*?\)\s*\{(.*?)\}\Z/m
-
-BRACE_SYNTAX = {
-  :evaluation_pattern    => /\{\{([\s\S]+?)\}\}/,
-  :interpolation_pattern => /\{\{=([\s\S]+?)\}\}/,
-  :escape_pattern        => /\{\{-([\s\S]+?)\}\}/,
-  :interpolation_with_subtemplate_pattern => %r{
-    \{\{=(?<start>(?:(?!\}\})[\s\S])+\{)\s*\}\}
-    (?<middle>
-    (
-      (?<re>
-        \{\{=?(?:(?!\}\})[\s\S])+\{\s*\}\}
-        (?:
-          \{\{\s*\}(?:(?!\{\s*\}\})[\s\S])+\{\s*\}\}
-          |
-          \g<re>
-          |
-          .{1}
-        )*?
-        \{\{\s*\}[^\{]+?\}\}
-      )
-      |
-      .{1}
-    )*?
-    )
-    \{\{\s*(?<end>\}[\s\S]+?)\}\}
-  }xm
-}
-
-QUESTION_MARK_SYNTAX = {
-  :evaluation_pattern    => /<\?([\s\S]+?)\?>/,
-  :interpolation_pattern => /<\?=([\s\S]+?)\?>/,
-  :escape_pattern        => /<\?-([\s\S]+?)\?>/,
-  :interpolation_with_subtemplate_pattern => %r{
-    <\?=(?<start>(?:(?!\?>)[\s\S])+\{)\s*\?>
-    (?<middle>
-    (
-      (?<re>
-        <\?=?(?:(?!\?>)[\s\S])+\{\s*\?>
-        (?:
-          <\?\s*\}(?:(?!\{\s*\?>)[\s\S])+\{\s*\?>
-          |
-          \g<re>
-          |
-          .{1}
-        )*?
-        <\?\s*\}[^\{]+?\?>
-      )
-      |
-      .{1}
-    )*?
-    )
-    <\?\s*(?<end>\}[\s\S]+?)\?>
-  }xm
-
-}
-
-
-class EJSCompilationTest < Minitest::Test
-  
-  test "compile" do
-    result = EJS.compile("Hello <%= name %>")
-    
-    assert_match FUNCTION_PATTERN, result
-    assert_no_match(/Hello \<%= name %\>/, result)
-  end
-
-  test "compile with custom syntax" do
-    standard_result = EJS.compile("Hello <%= name %>")
-    braced_result   = EJS.compile("Hello {{= name }}", BRACE_SYNTAX)
-
-    assert_match FUNCTION_PATTERN, braced_result
-    assert_equal standard_result, braced_result
-  end
-  
-end
-
-class EJSCustomPatternTest < Minitest::Test
-
-  def setup
-    @original_evaluation_pattern = EJS.evaluation_pattern
-    @original_interpolation_pattern = EJS.interpolation_pattern
-    EJS.evaluation_pattern = BRACE_SYNTAX[:evaluation_pattern]
-    EJS.interpolation_pattern = BRACE_SYNTAX[:interpolation_pattern]
-  end
-
-  def teardown
-    EJS.interpolation_pattern = @original_interpolation_pattern
-    EJS.evaluation_pattern = @original_evaluation_pattern
-  end
-
-  test "compile" do
-    result = EJS.compile("Hello {{= name }}")
-    assert_match FUNCTION_PATTERN, result
-    assert_no_match(/Hello \{\{= name \}\}/, result)
-  end
-
-  test "compile with custom syntax" do
-    standard_result = EJS.compile("Hello {{= name }}")
-    question_result = EJS.compile("Hello <?= name ?>", QUESTION_MARK_SYNTAX)
-
-    assert_match FUNCTION_PATTERN, question_result
-    assert_equal standard_result, question_result
-  end
-end
-
-class EJSEvaluationTest < Minitest::Test
+class EvaluationTest < Minitest::Test
 
   test "quotes" do
     template = "<%= thing %> is gettin' on my noives!"
